requirements.txt
asgiref==3.11.1
black==26.3.1
click==8.3.2
colorama==0.4.6
Django==6.0.4
django-rest-framework==0.1.0
djangorestframework==3.17.1
flake8==7.3.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.6
pycodestyle==2.14.0
pyflakes==3.4.0
pytokens==0.4.1
sqlparse==0.5.5
tzdata==2026.1
